Fix: Skip uid registry when metadata key value is 0
A value of 0 for the metadata key indicates that the metadata channel
was never created/pushed on the consumer side.

This can occur in scenario where a tracker(s) (vuid/vgid/vpid) is present.
The metadata channel might never be created/pushed since no applications
are actually tracing. Still, the uid registry exists due to the order
in which the sessiond received the client commands (create,
enable-channel, lttng track/untrack).

See included test for a complete reproducer.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>