Send ust and kernel domain directory handle to consumer
Send a directory handle to ust/ and kernel/ subdirectories of the
chunk to the consumer daemon, and alter the paths sent to consumer
so they do not include the domain subdir.

This will facilitate implementation of the clear command when no
rotation has been performed. The clear command will need to move the
ust/ and kernel/ directories to a "deleted" temporary directory while
the consumer is still actively using those directories. Making sure that
the consumer uses a handle on those subdirectories only ensures that the
consumer will not observe any transient state when those directories are
moved.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>