common trace-chunk: introduce lttng_trace_chunk_get_name_overridden
While it is possible to query whether the name of a chunk is
overridden through lttng_trace_chunk_get_name(), using it requires
a lot of code to deal with an unused name string when all that is needed
is to check whether the name is overridden.

In preparation for the new "path" handling needed for the clear
feature introduce this lttng_trace_chunk_get_name_overridden()
API to allow checking whether name is overridden in a succint way.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>