ust registry: Refactor representation of nested types
Those allow LTTng-UST to internally represent nested types properly, and
to serialize them when sending them over to the session daemon. However,
for now, the session daemon only accepts arrays and sequences which
contain an integer type, which is the only use-case emitted by
lttng-ust. Wait until we have the nested types wired up within
lttng-ust before accepting and supporting nested types so it can
be tested.

Given the size of this change to ust-metadata.c, use this opportunity
to make sure that every

  if (ret)
      ...

in this file is changed to conform to lttng-tools coding style:

  if (ret) {
      ...
  }

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I3ac44fc3a993f73d1fb08056331ad6fed7059981