tests: introduce --with-babeltrace-test-bin
Introduce --with-babeltrace-test-bin to allow specifying which
babeltrace binary should be used in tests.

If this option is not specified, use the babeltrace2 binary if found,
else fall back on babeltrace.

When looking for babeltrace2 and babeltrace, abide by
--with-babeltrace2-bin and --with-babeltrace-bin. If any of those are
not set, respectively use "babeltrace2" and "babeltrace".

Ensure all shell scripts and python tests use the BABELTRACE_BIN
environment variable.

Fixes #1270

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Co-developed-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I09006c4863df66f2f73980dce02ddc46e8f384ab